+++ client/my-sites/plans-features-main/index.tsx
@@ -34,7 +34,7 @@ import { isValidFeatureKey, FEATURES_LIST } from 'calypso/lib/plans/features-lis
 import scrollIntoViewport from 'calypso/lib/scroll-into-viewport';
 import PlanNotice from 'calypso/my-sites/plans-features-main/components/plan-notice';
 import PlanTypeSelector from 'calypso/my-sites/plans-features-main/components/plan-type-selector';
-import PlansGrid from 'calypso/my-sites/plans-grid';
+import { FeaturesGrid, ComparisonGrid } from 'calypso/my-sites/plans-grid';
 import useGridPlans from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-grid-plans';
 import usePlanFeaturesForGridPlans from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-plan-features-for-grid-plans';
 import useRestructuredPlanFeaturesForComparisonGrid from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-restructured-plan-features-for-comparison-grid';
@@ -46,6 +46,7 @@ import isEligibleForWpComMonthlyPlan from 'calypso/state/selectors/is-eligible-f
 import { getCurrentPlan } from 'calypso/state/sites/plans/selectors';
 import { getSitePlanSlug, getSiteSlug } from 'calypso/state/sites/selectors';
 import useAddOns from '../add-ons/hooks/use-add-ons';
+import ComparisonGridToggle from './components/comparison-grid-toggle';
 import { FreePlanFreeDomainDialog } from './components/free-plan-free-domain-dialog';
 import { FreePlanPaidDomainDialog } from './components/free-plan-paid-domain-dialog';
 import { LoadingPlaceHolder } from './components/loading-placeholder';
@@ -689,9 +690,8 @@ const PlansFeaturesMain = ( {
 						data-e2e-plans="wpcom"
 					>
 						<div className="plans-wrapper">
-							<PlansGrid
-								gridPlansForFeaturesGrid={ gridPlansForFeaturesGrid }
-								gridPlansForComparisonGrid={ gridPlansForComparisonGrid }
+							<FeaturesGrid
+								gridPlans={ gridPlansForFeaturesGrid }
 								gridPlanForSpotlight={ gridPlanForSpotlight }
 								paidDomainName={ paidDomainName }
 								wpcomFreeDomainSuggestion={ wpcomFreeDomainSuggestion }
@@ -726,6 +726,64 @@ const PlansFeaturesMain = ( {
 									} )
 								}
 							/>
+							{ ! hidePlansFeatureComparison && (
+								<ComparisonGridToggle
+									onClick={ toggleShowPlansComparisonGrid }
+									label={
+										showPlansComparisonGrid
+											? translate( 'Hide comparison' )
+											: translate( 'Compare plans' )
+									}
+									ref={ observableForOdieRef }
+								/>
+							) }
+							{ ! hidePlansFeatureComparison && showPlansComparisonGrid ? (
+								<div
+									ref={ plansComparisonGridRef }
+									className="plans-features-main__comparison-grid-container"
+								>
+									<ComparisonGrid
+										gridPlans={ gridPlansForComparisonGrid }
+										gridPlanForSpotlight={ gridPlanForSpotlight }
+										paidDomainName={ paidDomainName }
+										wpcomFreeDomainSuggestion={ wpcomFreeDomainSuggestion }
+										isCustomDomainAllowedOnFreePlan={ isCustomDomainAllowedOnFreePlan }
+										isInSignup={ isInSignup }
+										isLaunchPage={ isLaunchPage }
+										onUpgradeClick={ handleUpgradeClick }
+										flowName={ flowName }
+										selectedFeature={ selectedFeature }
+										selectedPlan={ selectedPlan }
+										siteId={ siteId }
+										isReskinned={ isReskinned }
+										intervalType={ intervalType }
+										hidePlansFeatureComparison={ hidePlansFeatureComparison }
+										hideUnavailableFeatures={ hideUnavailableFeatures }
+										currentSitePlanSlug={ sitePlanSlug }
+										planActionOverrides={ planActionOverrides }
+										intent={ intent }
+										showLegacyStorageFeature={ showLegacyStorageFeature }
+										showUpgradeableStorage={ showUpgradeableStorage }
+										stickyRowOffset={ masterbarHeight }
+										usePricingMetaForGridPlans={ usePricingMetaForGridPlans }
+										allFeaturesList={ FEATURES_LIST }
+										showPlansComparisonGrid={ showPlansComparisonGrid }
+										toggleShowPlansComparisonGrid={ toggleShowPlansComparisonGrid }
+										planTypeSelectorProps={ planTypeSelectorProps }
+										ref={ plansComparisonGridRef }
+										observableForOdieRef={ observableForOdieRef }
+										onStorageAddOnClick={ ( addOnSlug ) =>
+											recordTracksEvent( 'calypso_signup_storage_add_on_dropdown_option_click', {
+												add_on_slug: addOnSlug,
+											} )
+										}
+									/>
+									<ComparisonGridToggle
+										onClick={ toggleShowPlansComparisonGrid }
+										label={ translate( 'Hide comparison' ) }
+									/>
+								</div>
+							) : null }
 						</div>
 					</div>
 				</>
