+++ client/my-sites/plans-grid/style.scss
@@ -1054,18 +1054,18 @@ body.is-section-signup.is-white-signup,
  * However, on the /plans page, due to the presence of the masterbar on top,
  * we need to add a scroll offset.
  */
-.is-section-plans .plan-features-2023-grid__plan-comparison-grid-container {
+.is-section-plans .plans-features-main__comparison-grid-container {
 	scroll-margin-top: $plan-features-header-banner-height + 24px;
 	margin: auto;
 	@include plan-comparison-grid-breakpoints-with-sidebar;
 }
 
-.is-section-signup .plan-features-2023-grid__plan-comparison-grid-container {
+.is-section-signup .plans-features-main__comparison-grid-container {
 	margin: auto;
 	@include plan-comparison-grid-breakpoints;
 }
 
-.is-section-stepper .plan-features-2023-grid__plan-comparison-grid-container {
+.is-section-stepper .plans-features-main__comparison-grid-container {
 	margin: auto;
 	@include plan-comparison-grid-breakpoints;
 }
