+++ client/my-sites/plans/ecommerce-trial/wooexpress-plans/style.scss
@@ -25,12 +25,12 @@
 }
 
 .wooexpress-plans__grid.is-2023-pricing-grid {
-	.plan-features-2023-grid__header-title {
+	.plans-grid__header-title {
 		font-family: $sans;
 	}
 
-	.plan-features-2023-grid__content .plan-features-2023-grid__table-item .plan-price,
-	.plan-features-2023-grid__plan-comparison-grid-container .plan-comparison-grid .plan-price {
+	.plans-grid__content .plans-grid__table-item .plan-price,
+	.plans-grid__plan-comparison-grid-container .plan-comparison-grid .plan-price {
 		font-family: $sans;
 	}
 
@@ -38,11 +38,11 @@
 		font-size: $woo-font-title-medium;
 	}
 
-	.plan-features-2023-grid__plan-comparison-grid-container .plan-comparison-grid .wp-brand-font {
+	.plans-grid__plan-comparison-grid-container .plan-comparison-grid .wp-brand-font {
 		display: none;
 	}
 
-	.plan-features-2023-grid__table.has-2-cols {
+	.plans-grid__table.has-2-cols {
 		margin-left: auto;
 		margin-right: auto;
 		max-width: 800px;
